requirements.txt
Pillow==11.2.1
pix2tex==0.1.4
pyperclip==1.9.0
